refactor: reorganize http handlers, imports and regenerate wire di :hammer:
