+++ components/ledger/internal/ports/http/ledger.go
@@ -1,4 +1,4 @@
-package ports
+package http
 
 import (
 	"os"
