+++ components/ledger/internal/ports/http/organization.go
@@ -1,4 +1,4 @@
-package ports
+package http
 
 import (
 	"os"
