+++ components/ledger/internal/ports/http/product.go
@@ -1,4 +1,4 @@
-package ports
+package http
 
 import (
 	"github.com/LerianStudio/midaz/common/mlog"
